pyproject.toml
[project]
name = "solc-event-extractor"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Evan K", email = "ekcopersonal@gmail.com" }]
dependencies = ["py-solc-x>=2.0.3"]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/solc_event_extractor"]
